+++ HISTORY.md
@@ -10,3 +10,13 @@ Initial release!
 ------------------
 
 Python 3 compatibility
+
+0.3.0 (2018-02-28)
+------------------
+
+New bulk handler + event cleanup.
+
+* Added `LogglyBulkHandler`
+* `LogglyHandler` uses own instance of `ThreadPoolExecutor`
+* Register `atexit` cleanup methods
+* `.flush` methods for both handlers to explicitly flush pending events